Criteria for ElemElem_isAdjacent is updated
In this version, the face list "FaceReflist" round a fixed node (derived from N[E]) is used as reference to check whether two elements are connected. If two elements are not connected by a face which belong to FaceReflist, then they are regarded separated, even they may be adjacent by some other faces.